tests: add Postgres persistence integration tests (closes #65)
E2E tests using testcontainers for Postgres repo layer:
- apply_spec → create_run → update_status → list_runs/latest → record/list_artifacts
- persistence verified via repo restart
