+++ apps/control-plane/Cargo.toml.bak
@@ -28,9 +28,13 @@ astra-core = { path = "../../crates/astra-core" }
 astra-metadata = { path = "../../crates/astra-metadata" }
 astra-observability = { path = "../../crates/astra-observability" }
 astra-secrets = { path = "../../crates/astra-secrets" }
-astra-yaml = { path = "../../crates/astra-yaml" }
+astra-yaml = { path = "../../crates/astra-yaml", default-features = false }
 
 [dev-dependencies]
+astra-yaml = { path = "../../crates/astra-yaml" }
+tokio = { workspace = true, features = ["full"] }
+
 testcontainers = "0.15"
 tokio-postgres = "0.7"
-serde_yaml.workspace = true
+uuid.workspace = true
+serde_json.workspace = true
