+++ apps/control-plane/src/lib.rs
@@ -1,10 +1,19 @@
-pub mod repositories;
-pub mod services;
+pub mod api;
+pub mod app;
+pub mod config;
+pub mod error;
 pub mod http;
+pub mod metadata;
 pub mod models;
+pub mod repositories;
+pub mod scheduler;
+pub mod services;
+pub mod state;
+
+pub use astra_yaml::AstraSpec;
 
 pub use repositories::{
-    AppliedPipelineRecord, CreatePipelineRunRecord, PipelineRecord, PipelineRepository,
-    PipelineRunRecord, RecordStagedArtifactRecord, StagedArtifactRecord,
+    AppliedPipelineRecord, CreatePipelineRunRecord, InMemoryPipelineRepository, PipelineRecord,
+    PipelineRepository, PipelineRunRecord, PostgresPipelineRepository, RecordStagedArtifactRecord,
+    StagedArtifactRecord,
 };
-pub use astra_yaml::AstraSpec;
